// src/entities/ITEMS/phone-item/types/phone-item.types.ts
export type PhoneSubNames =
    | ""
    | "Apple"
    | "Samsung"
    | "Xiaomi"
    | "OPPO"
    | "Realme"
    | "OnePlus"

export type PhoneTypes = {
    id: number
    name: string
    price: number
    subName: string
    RAM: string
    gigabyte: number
    "batteryCapacity(mAh)": number
    equipment: PhoneEquipment
    defects: PhoneDefects
    rechargeCycles: PhoneRechargeCycle
}

export type PhoneEquipment = {
    ref: number
    fullEquipment: number
    withoutEquipment: number
    originalBox: number
    emptyBox: number
    "eSIM(noPhysicalSIMslot)": number
    "noStylus(suppliedFactory)": number
    isCharger: number
    isHeadphones: number
    isCover: number
}

export type PhoneDefects = {
    caseWasOpened: number
    caseReplacedCase: number
    bodyHasStrongDentCracks: number
    bodyHasDentChips: number
    caseScratchesBackCover: number
    "TouchID(home)": number
    "FaceID(broken)": number
    displayHasDeepScratches: number
    displaySpotStripes: number
    "displayCrack(replacementRequired)": number
    displayReplaced: number
    cameraSpotShooting: number
    cameraReplaceable: number
    cameraCrackGlass: number
    cameraNotWorking: number
    "speakerTop(auditory)": number
    "speakerLower(conversational)": number
    "sideButtons(replacement)": number
    vibrationDoesntWork: number
    flashlightDoesntWork: number
}